Throw for navigation to and keys on query types. Add more test infrastructure for query types.
Fixes #13662
