+++ test/EFCore.Relational.Tests/RelationalModelValidatorTest.cs
@@ -18,7 +18,7 @@ using Xunit;
 // ReSharper disable InconsistentNaming
 namespace Microsoft.EntityFrameworkCore
 {
-    public class RelationalModelValidatorTest : ModelValidatorTest
+    public class RelationalModelValidatorTest : ModelValidatorTestBase
     {
         [Fact]
         public virtual void Ignores_bool_with_default_value_false()
