+++ test/EFCore.Tests/ModelBuilding/ModelBuilderNonGenericTest.cs
@@ -78,11 +78,18 @@ namespace Microsoft.EntityFrameworkCore.ModelBuilding
                 return this;
             }
 
+            public override TestOwnedEntityTypeBuilder<TEntity> Owned<TEntity>()
+                => new NonGenericTestOwnedEntityTypeBuilder<TEntity>(ModelBuilder.Owned(typeof(TEntity)));
+
             public override TestQueryTypeBuilder<TQuery> Query<TQuery>()
                 => new NonGenericTestQueryTypeBuilder<TQuery>(ModelBuilder.Query(typeof(TQuery)));
 
-            public override TestOwnedEntityTypeBuilder<TEntity> Owned<TEntity>()
-                => new NonGenericTestOwnedEntityTypeBuilder<TEntity>(ModelBuilder.Owned(typeof(TEntity)));
+            public override TestModelBuilder Query<TQuery>(Action<TestQueryTypeBuilder<TQuery>> buildAction)
+            {
+                ModelBuilder.Query(
+                    typeof(TQuery), queryTypeBuilder => buildAction(new NonGenericTestQueryTypeBuilder<TQuery>(queryTypeBuilder)));
+                return this;
+            }
 
             public override TestModelBuilder Ignore<TEntity>()
             {
